fix: correct callback return type from uint64 to uint8 in custom action and recognition

+++ custom_action.go
@@ -59,7 +59,7 @@ func _MaaCustomActionCallbackAgent(
 	recoId int64,
 	box uintptr,
 	transArg uintptr,
-) uint64 {
+) uint8 {
 	// Here, we are simply passing the uint64 value as a pointer
 	// and will not actually dereference this pointer.
 	id := uint64(transArg)
