+++ custom_recognition.go
@@ -65,7 +65,7 @@ func _MaaCustomRecognitionCallbackAgent(
 	image, roi uintptr,
 	transArg uintptr,
 	outBox, outDetail uintptr,
-) uint64 {
+) uint8 {
 	// Here, we are simply passing the uint64 value as a pointer
 	// and will not actually dereference this pointer.
 	id := uint64(transArg)
